requirements.txt
langchain>=1.0.5
langchain-core>=1.0.4
langchain-openai>=1.0.2
langchain-community>=0.4.1
langchain-text-splitters>=1.0.0

# LangGraph v1.0 (upgraded from v0.x)
langgraph>=1.0.3
langgraph-prebuilt>=1.0.2

# Deep Agents (built on LangGraph for complex multi-step tasks)
deepagents

# Environment variable management
python-dotenv>=1.0.0

# Vision & Inference dependencies (for inference.py)
torch>=2.0.0
transformers>=4.30.0
pillow>=9.0.0
accelerate>=0.20.0  # For model loading optimization